requirements.txt
tensorflow
numpy
pandas
seaborn
matplotlib
Flask
pyYAML
omegaconf
-e .